+++ packages/google_sign_in/google_sign_in_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+
 ## 0.12.4+4
 
 * Asserts that new `forceAccountName` parameter is null (not used in web).