+++ packages/vector_graphics_codec/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+
 ## 1.1.13
 
 * Works around a subtle Wasm bug in `writeRadialGradient`.